:recycle: Reorganize and use static index page
