+++ lib/models/ErrorTypes.js
