--- src/components/InlineCode.js
@@ -1,7 +1,0 @@
-const html = require('../utils/html')
-
-module.exports = function InlineCode({ children }) {
-  return html`
-    <code class="dib code">${children}</code>
-  `
-}
